# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WhatsApp Attendant API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface WhatsappState {
  /** disconnected | qr_pending | connecting | connected | logged_out */
  status: string;
  /**
   * Base64 PNG data URL of the pairing QR code
   * @nullable
   */
  qrCode: string | null;
  /** @nullable */
  phoneNumber: string | null;
  /** @nullable */
  connectedAt: string | null;
}

export interface Settings {
  /** agent (AI auto-replies) | owner (user is online and AI is paused) */
  mode: string;
  ownerName: string;
  agentName: string;
  agentPersona: string;
  voiceEnabled: boolean;
  voice: string;
  autoReplyAudioForAudio: boolean;
}

export interface UpdateSettingsBody {
  mode?: string;
  ownerName?: string;
  agentName?: string;
  agentPersona?: string;
  voiceEnabled?: boolean;
  voice?: string;
  autoReplyAudioForAudio?: boolean;
}

export interface Contact {
  jid: string;
  name: string;
  /** @nullable */
  lastMessageAt: string | null;
  /** @nullable */
  lastMessagePreview: string | null;
  unread: number;
  messageCount: number;
}

export interface Message {
  id: number;
  contactJid: string;
  fromMe: boolean;
  isAi: boolean;
  /** text | audio */
  kind: string;
  /** @nullable */
  text: string | null;
  /** @nullable */
  audioUrl: string | null;
  /** @nullable */
  durationSeconds: number | null;
  createdAt: string;
}

export interface SendMessageBody {
  text: string;
}

export interface Stats {
  totalContacts: number;
  totalMessages: number;
  aiRepliesToday: number;
  manualRepliesToday: number;
  incomingToday: number;
  connected: boolean;
  mode: string;
}