# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: WhatsApp Attendant API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /contacts/{jid}/messages:
    get:
      operationId: listContactMessages
      tags: [messages]
      summary: List messages for a contact
      parameters:
        - in: path
          name: jid
          required: true
          schema:
            type: string
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendContactMessage
      tags: [messages]
      summary: Send a message manually as the owner
      parameters:
        - in: path
          name: jid
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageBody"
      responses:
        "201":
          description: ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
...
components:
  schemas:
...
    Message:
      type: object
      properties:
        id:
          type: integer
        contactJid:
          type: string
        fromMe:
          type: boolean
        isAi:
          type: boolean
        kind:
          type: string
          description: text | audio
        text:
          type: ["string", "null"]
        audioUrl:
          type: ["string", "null"]
        durationSeconds:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - contactJid
        - fromMe
        - isAi
        - kind
        - text
        - audioUrl
        - durationSeconds
        - createdAt
    SendMessageBody:
      type: object
      properties:
        text:
          type: string
      required:
        - text
...