# lib/api-zod/src/generated/types/message.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WhatsApp Attendant API
 * OpenAPI spec version: 0.1.0
 */

export interface Message {
  id: number;
  contactJid: string;
  fromMe: boolean;
  isAi: boolean;
  /** text | audio */
  kind: string;
  /** @nullable */
  text: string | null;
  /** @nullable */
  audioUrl: string | null;
  /** @nullable */
  durationSeconds: number | null;
  createdAt: Date;
}